+++ src/main/scala/AnyOps.scala
@@ -20,4 +20,5 @@ object AnyOps {
     @targetName("notEquals")
     def !==(other: A): Boolean = !(self === other)
   }
+
 }