+++ src/test/resources/users.sql
@@ -1,4 +1,4 @@
-CREATE TABLE public.administration (
+CREATE TABLE public.administrators (
     username character varying(100) NOT NULL,
     password character varying(60) NOT NULL
 );
@@ -17,8 +17,12 @@ CREATE TABLE public.store_managers (
     store integer NOT NULL
 );
 
-ALTER TABLE ONLY public.administration ADD CONSTRAINT administration_pkey PRIMARY KEY (username);
+ALTER TABLE ONLY public.administrators ADD CONSTRAINT administrators_pkey PRIMARY KEY (username);
 
 ALTER TABLE ONLY public.customers ADD CONSTRAINT customers_pkey PRIMARY KEY (email);
 
 ALTER TABLE ONLY public.store_managers ADD CONSTRAINT store_managers_pkey PRIMARY KEY (username);
+
+INSERT INTO public.administrators(
+  username, password)
+  VALUES ('elena', 'Efda!dWQ');