+++ users.sql
@@ -1,4 +1,4 @@
-CREATE TABLE public.administration (
+CREATE TABLE public.administrators (
     username character varying(100) NOT NULL,
     password character varying(60) NOT NULL
 );
@@ -17,7 +17,7 @@ CREATE TABLE public.store_managers (
     store integer NOT NULL
 );
 
-ALTER TABLE ONLY public.administration ADD CONSTRAINT administration_pkey PRIMARY KEY (username);
+ALTER TABLE ONLY public.administrators ADD CONSTRAINT administrators_pkey PRIMARY KEY (username);
 
 ALTER TABLE ONLY public.customers ADD CONSTRAINT customers_pkey PRIMARY KEY (email);